src/main/java/com/codingbat/array1/Front11.java
/*
  CodingBat Java Array 1 - https://codingbat.com/java/Array-1

  My repo - https://github.com/egalli64/jex/
 */
package com.codingbat.array1;

/**
 * front11 - https://codingbat.com/prob/p128270
 * <p>
 * Given two arrays, generate a new array containing the first elements of the
 * two input ones. When an input array is empty, ignore it.
 */
public class Front11 {
    /**
     * Check the arrays in before creating the array out
     * 
     * @param a first possibly empty array
     * @param b second possibly empty array
     * @return the first elements of the arrays (when available)
     */
    public static int[] solution(int[] a, int[] b) {
        boolean aGood = a.length > 0;
        boolean bGood = b.length > 0;
        int[] result = new int[(aGood ? 1 : 0) + (bGood ? 1 : 0)];

        if (aGood) {
            result[0] = a[0];
        }
        if (bGood) {
            result[aGood ? 1 : 0] = b[0];
        }

        return result;
    }
}
